+++ service/internal/cloudconfig/tccpn_extension.go
@@ -218,7 +218,33 @@ func (e *TCCPNExtension) Files() ([]k8scloudconfig.FileAsset, error) {
 			},
 			Permissions: 0744,
 		}
-		filesMeta = append(filesMeta, etcdClusterMigratorManifest, etcdClusterMigratorInstaller, healthChecker)
+		lifeCycleContinue := k8scloudconfig.FileMetadata{
+			AssetContent: template.MasterInstanceLifecycleContinue,
+			Path:         "/opt/bin/master-instance-lifecycle-continue",
+			Owner: k8scloudconfig.Owner{
+				Group: k8scloudconfig.Group{
+					Name: FileOwnerGroupName,
+				},
+				User: k8scloudconfig.User{
+					Name: FileOwnerUserName,
+				},
+			},
+			Permissions: 0744,
+		}
+		etcdHealthCheck := k8scloudconfig.FileMetadata{
+			AssetContent: template.ETCDHealthCheck,
+			Path:         "/opt/bin/etcd-healthcheck",
+			Owner: k8scloudconfig.Owner{
+				Group: k8scloudconfig.Group{
+					Name: FileOwnerGroupName,
+				},
+				User: k8scloudconfig.User{
+					Name: FileOwnerUserName,
+				},
+			},
+			Permissions: 0744,
+		}
+		filesMeta = append(filesMeta, etcdClusterMigratorManifest, etcdClusterMigratorInstaller, healthChecker, lifeCycleContinue, etcdHealthCheck)
 	}
 
 	if !e.hasCilium {
@@ -464,7 +490,13 @@ func (e *TCCPNExtension) Units() ([]k8scloudconfig.UnitAsset, error) {
 			Name:         "master-instance-healthcheck.timer",
 			Enabled:      true,
 		}
-		unitsMeta = append(unitsMeta, etcdClusterMigratorService, healthCheckService, healthCheckTimer)
+		// Lifecycle hook after etcd started.
+		lifeCycleContinue := k8scloudconfig.UnitMetadata{
+			AssetContent: template.MasterInstanceLifecycleContinueService,
+			Name:         "master-instance-lifecycle-continue.service",
+			Enabled:      true,
+		}
+		unitsMeta = append(unitsMeta, etcdClusterMigratorService, healthCheckService, healthCheckTimer, lifeCycleContinue)
 	}
 
 	var newUnits []k8scloudconfig.UnitAsset